+++ common/instructions/condition.h
@@ -8,7 +8,7 @@ class Condition : public Instruction
 public:
     static const std::string MNEMONIC;
     Condition(uint16_t input_memory_address, uint8_t subprogram_index);
-    void operator()(std::vector<float>& memory, uint8_t& subprogram_index, std::vector<uint16_t>& instruction_addresses,
+    void operator()(std::vector<int32_t>& memory, uint8_t& subprogram_index, std::vector<uint16_t>& instruction_addresses,
         std::vector<uint8_t>& return_indices) const override;
     std::vector<std::string> toStringTokens() const override;
     static std::unique_ptr<Instruction> fromStringTokens(const std::vector<std::string>& tokens);
