+++ common/instructions/negate.h
@@ -8,7 +8,7 @@ class Negate : public Instruction
 public:
     static const std::string MNEMONIC;
     Negate(uint16_t input_memory_address, uint16_t output_memory_address);
-    void operator()(std::vector<float>& memory, uint8_t& subprogram_index, std::vector<uint16_t>& instruction_addresses,
+    void operator()(std::vector<int32_t>& memory, uint8_t& subprogram_index, std::vector<uint16_t>& instruction_addresses,
         std::vector<uint8_t>& return_indices) const override;
     std::vector<std::string> toStringTokens() const override;
     static std::unique_ptr<Instruction> fromStringTokens(const std::vector<std::string>& tokens);
